fundus-vision/requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.9
albumentations==1.4.7
opencv-python-headless>=4.10.0.84
pillow>=10.4.0
numpy>=1.26
pandas>=2.2.2
scikit-learn>=1.4.2
flask>=3.0.0
gunicorn>=21.2.0
pyyaml>=6.0.1
